Verified property types with Henrique. Ready for final review.
