+++ mcs_benchmark_data/models/benchmark_context.py
@@ -3,7 +3,7 @@ from dataclasses_json import LetterCase, dataclass_json
 
 from rdflib import Graph, URIRef
 from rdflib.resource import Resource
-from mcs_benchmark_data.namespace import RDF, MCS
+from mcs_benchmark_data.namespace import RDF, MCS, SCHEMA
 
 from mcs_benchmark_data._model import _Model
 
@@ -18,7 +18,7 @@ class BenchmarkContext(_Model):
 
     def to_rdf(self, *, graph: Graph) -> Resource:
         resource = _Model.to_rdf(self, graph=graph)
-        graph.add((self.antecedent_uri, MCS.context, self.uri))
-        resource.add(MCS.text, self._quote_rdf_literal(self.text))
+        graph.add((self.antecedent_uri, MCS.BenchmarkContext, self.uri))
+        resource.add(SCHEMA.text, self._quote_rdf_literal(self.text))
 
         return resource
