+++ mcs_benchmark_data/models/benchmark_prompt.py
@@ -5,7 +5,7 @@ from rdflib import Graph, URIRef
 from rdflib.resource import Resource
 
 from mcs_benchmark_data._model import _Model
-from mcs_benchmark_data.namespace import MCS
+from mcs_benchmark_data.namespace import MCS, SCHEMA
 
 
 @dataclass_json(letter_case=LetterCase.CAMEL)
@@ -21,6 +21,6 @@ class BenchmarkPrompt(_Model):
     def to_rdf(self, *, graph: Graph) -> Resource:
         resource = _Model.to_rdf(self, graph=graph)
         graph.add((self.antecedent_uri, MCS.prompt, self.uri))
-        resource.add(MCS.text, self._quote_rdf_literal(self.text))
+        resource.add(SCHEMA.text, self._quote_rdf_literal(self.text))
 
         return resource
