+++ mcs_benchmark_data/models/benchmark_solution.py
@@ -1,11 +1,6 @@
 from dataclasses import dataclass
 from dataclasses_json import LetterCase, dataclass_json
 
-from mcs_benchmark_data.namespace import MCS, RDF
-from rdflib import Graph
-from rdflib.resource import Resource
-
-from mcs_benchmark_data._model import _Model
 from mcs_benchmark_data.models.benchmark_choice import BenchmarkChoice
 
 
@@ -15,9 +10,3 @@ class BenchmarkSolution(BenchmarkChoice):
     """A benchmark's sample solution choice"""
 
     """Choice of Physical IQA (goals)"""
-
-    def to_rdf(self, *, graph: Graph) -> Resource:
-        resource = _Model.to_rdf(self, graph=graph)
-        resource.add(RDF.type, MCS.BenchmarkChoice)
-
-        return resource
