+++ mcs_benchmark_data/models/dev_score.py
@@ -4,7 +4,7 @@ from dataclasses_json import LetterCase, dataclass_json
 from rdflib import Graph, URIRef
 from rdflib.resource import Resource
 
-from mcs_benchmark_data.namespace import XSD, MCS
+from mcs_benchmark_data.namespace import MCS, SCHEMA
 
 from mcs_benchmark_data._model import _Model
 
@@ -22,8 +22,8 @@ class DevScore(_Model):
     def to_rdf(self, *, graph: Graph) -> Resource:
         resource = _Model.to_rdf(self, graph=graph)
         graph.add((self.submission_uri, MCS.devScore, self.uri))
-        resource.add(MCS.isBasedOn, self._quote_rdf_literal(self.is_based_on))
-        resource.add(MCS.name, self._quote_rdf_literal(self.name))
-        resource.add(MCS.value, self._quote_rdf_literal(self.value))
+        resource.add(SCHEMA.isBasedOn, self._quote_rdf_literal(self.is_based_on))
+        resource.add(SCHEMA.name, self._quote_rdf_literal(self.name))
+        resource.add(SCHEMA.value, self._quote_rdf_literal(self.value))
 
         return resource
